requirements.txt
# NAIADS deps
-e git+https://royarzun@git.earthdata.nasa.gov/scm/naiads/xmsg-python.git#egg=xmsg
-e git+https://royarzun@git.earthdata.nasa.gov/scm/naiads/clara-python.git#egg=clara

# Django deps
Django==1.7.7
djangorestframework==3.1.1
djangorestframework-yaml==1.0.2
django-rest-swagger==0.2.9
PyYAML==3.11
simplejson==3.8.1

# Time based database
influxdb==2.12.0

# Test deps
mock==1.3.0
coverage==3.7.1
mockito>=0.6.1

# xMsg deps
pyzmq>=14.5.0
protobuf>=2.6
enum34>=1.0.4
argparse>=1.2.1
netifaces>=0.10.4
pytest>=2.9.2

# Clara deps
psutil==4.1.0
ordered-set==2.0.1
bitstring==3.1.4
